Throw IOException and expose simple delete() method on Deleter
This is the prime usage for IOException, so let's use it. Also expose another method that tries to delete a single file.

+++ subprojects/execution/src/main/java/org/gradle/internal/execution/impl/OutputsCleaner.java
@@ -91,7 +91,11 @@ public class OutputsCleaner {
     }
 
     public void cleanupDirectories() throws IOException {
-        for (File directory = directoriesToDelete.poll(); directory != null; directory = directoriesToDelete.poll()) {
+        while (true) {
+            File directory = directoriesToDelete.poll();
+            if (directory == null) {
+                break;
+            }
             if (isEmpty(directory)) {
                 LOGGER.debug("Deleting stale empty output directory '{}'.", directory);
                 Files.delete(directory.toPath());
